# 2-2-t/specs/001-smartbooking-scheduling/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: SmartBooking API
  version: 0.1.0
  description: |
    Contract-first REST API for SmartBooking.

    Notes:
    - Auth: JWT Bearer token in Authorization header.
    - All errors use a unified envelope with stable error codes and requestId.
servers:
  - url: http://localhost:3001
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new account (USER or PROVIDER)
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout (stateless token; client drops token)
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/password/forgot:
    post:
      tags: [Auth]
      summary: Request a password reset token
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '202':
          description: Accepted (always returns success to avoid user enumeration)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/auth/password/reset:
    post:
      tags: [Auth]
      summary: Reset password with token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/bookings:
    get:
      tags: [Booking]
      summary: List my bookings
      operationId: listMyBookings
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/BookingStatus'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'
                required: [items]
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Booking]
      summary: Create a booking (reserve a seat)
      operationId: createBooking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/bookings/{bookingId}/cancel:
    post:
      tags: [Booking]
      summary: Cancel my booking (deadline enforced)
      operationId: cancelBooking
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/bookingId'
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/provider/bookings:
    get:
      tags: [Provider]
      summary: List bookings for services owned by current provider
      operationId: providerListBookings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProviderBooking'
                required: [items]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/provider/bookings/{bookingId}/status:
    post:
      tags: [Provider]
      summary: Update booking status (state machine enforced)
      operationId: providerUpdateBookingStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/bookingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBookingStatusRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List users
      operationId: adminListUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                required: [items]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/users/{userId}/suspend:
    post:
      tags: [Admin]
      summary: Suspend a user
      operationId: adminSuspendUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/users/{userId}/activate:
    post:
      tags: [Admin]
      summary: Activate a user
      operationId: adminActivateUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/services:
    get:
      tags: [Admin]
      summary: List services (admin)
      operationId: adminListServices
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
                required: [items]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/services/{serviceId}/activate:
    post:
      tags: [Admin]
      summary: Activate a service
      operationId: adminActivateService
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serviceId'
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/services/{serviceId}/inactivate:
    post:
      tags: [Admin]
      summary: Inactivate a service
      operationId: adminInactivateService
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serviceId'
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    serviceId:
      name: serviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    bookingId:
      name: bookingId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [USER, PROVIDER]
      required: [email, password, role]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 8
      required: [token, newPassword]
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        user:
          $ref: '#/components/schemas/MeResponse'
      required: [accessToken, user]
    MeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [USER, PROVIDER, ADMIN]
        status:
          type: string
          enum: [ACTIVE, SUSPENDED]
      required: [id, email, role, status]
...
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        providerId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        createdAt:
          type: string
          format: date-time
      required: [id, providerId, name, description, durationMinutes, status, createdAt]
...
    UpdateBookingStatusRequest:
      type: object
      properties:
        targetStatus:
          type: string
          enum: [CONFIRMED, COMPLETED, CANCELLED]
      required: [targetStatus]
    ProviderBooking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        userEmail:
          type: string
          format: email
        serviceId:
          type: string
          format: uuid
        serviceName:
          type: string
        timeSlotId:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        cancelDeadlineAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/BookingStatus'
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userEmail
        - serviceId
        - serviceName
        - timeSlotId
        - startTime
        - endTime
        - cancelDeadlineAt
        - status
        - createdAt
...
    CreateBookingRequest:
      type: object
      properties:
        timeSlotId:
          type: string
          format: uuid
      required: [timeSlotId]
    BookingStatus:
      type: string
      enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED]
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        timeSlotId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/BookingStatus'
        createdAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, userId, timeSlotId, status, createdAt]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [USER, PROVIDER, ADMIN]
        status:
          type: string
          enum: [ACTIVE, SUSPENDED]
        createdAt:
          type: string
          format: date-time
      required: [id, email, role, status, createdAt]
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Stable internal error code
              example: BOOKING_SEAT_FULL
            message:
              type: string
              description: Human-readable error message
            requestId:
              type: string
              description: Correlation id for support/debugging
          required: [code, message, requestId]
      required: [error]